mcp/mcp-server/requirements.txt
# MCP Server dependencies
mcp>=1.0.0

# HTTP Server dependencies
fastapi>=0.104.1
uvicorn>=0.24.0

# Environment management
python-dotenv>=1.0.0

# RAG dependencies (inherited from parent project)
langchain>=0.3.7
langchain-openai>=0.2.8
langchain-community>=0.3.7
qdrant-client>=1.7.0
openpyxl>=3.1.2
requests>=2.31.0

# SQL database dependencies
pandas>=2.2.2
psycopg2-binary>=2.9.9
